@@ -11,15 +11,16 @@
 
 namespace spec\Memio\Validator\ModelValidator;
 
-use Memio\Validator\ViolationCollection;
 use Memio\Model\Argument;
+use Memio\Validator\ModelValidator;
+use Memio\Validator\ViolationCollection;
 use PhpSpec\ObjectBehavior;
 
 class ArgumentValidatorSpec extends ObjectBehavior
 {
     function it_is_a_model_validator()
     {
-        $this->shouldImplement('Memio\Validator\ModelValidator');
+        $this->shouldImplement(ModelValidator::class);
     }
 
     function it_supports_arguments(Argument $model)
