@@ -15,37 +15,38 @@ use Memio\Model\Argument;
 use Memio\Model\Constant;
 use Memio\Model\Method;
 use Memio\Model\Property;
+use Memio\Validator\ModelValidator;
 use PhpSpec\ObjectBehavior;
 
 class CollectionValidatorSpec extends ObjectBehavior
 {
     function it_is_a_model_validator()
     {
-        $this->shouldImplement('Memio\Validator\ModelValidator');
+        $this->shouldImplement(ModelValidator::class);
     }
 
     function it_supports_argument_collections(Argument $model)
     {
-        $this->supports(array($model))->shouldBe(true);
+        $this->supports([$model])->shouldBe(true);
     }
 
     function it_supports_constant_collections(Constant $model)
     {
-        $this->supports(array($model))->shouldBe(true);
+        $this->supports([$model])->shouldBe(true);
     }
 
     function it_supports_method_collections(Method $model)
     {
-        $this->supports(array($model))->shouldBe(true);
+        $this->supports([$model])->shouldBe(true);
     }
 
     function it_supports_property_collections(Property $model)
     {
-        $this->supports(array($model))->shouldBe(true);
+        $this->supports([$model])->shouldBe(true);
     }
 
     function it_does_not_support_empty_collections()
     {
-        $this->supports(array())->shouldBe(false);
+        $this->supports([])->shouldBe(false);
     }
 }
