@@ -11,6 +11,7 @@
 
 namespace spec\Memio\Validator\Violation;
 
+use Memio\Validator\Violation;
 use PhpSpec\ObjectBehavior;
 
 class SomeViolationSpec extends ObjectBehavior
@@ -24,7 +25,7 @@ class SomeViolationSpec extends ObjectBehavior
 
     function it_is_a_violation()
     {
-        $this->shouldHaveType('Memio\Validator\Violation');
+        $this->shouldHaveType(Violation::class);
     }
 
     function it_has_a_message()
