@@ -12,16 +12,15 @@
 namespace Memio\Validator\ModelValidator;
 
 use Memio\Model\Argument;
-use Memio\Validator\Constraint;
-use Memio\Validator\ConstraintValidator;
-use Memio\Validator\ModelValidator;
-use Memio\Validator\ViolationCollection;
+use Memio\Validator\{
+    Constraint,
+    ConstraintValidator,
+    ModelValidator,
+    ViolationCollection
+};
 
 class ArgumentValidator implements ModelValidator
 {
-    /**
-     * @var ConstraintValidator
-     */
     private $constraintValidator;
 
     public function __construct()
@@ -29,26 +28,17 @@ class ArgumentValidator implements ModelValidator
         $this->constraintValidator = new ConstraintValidator();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function add(Constraint $constraint)
     {
         $this->constraintValidator->add($constraint);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($model)
+    public function supports($model) : bool
     {
         return $model instanceof Argument;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function validate($model)
+    public function validate($model) : ViolationCollection
     {
         if (!$this->supports($model)) {
             return new ViolationCollection();
