@@ -11,30 +11,23 @@
 
 namespace Memio\Validator\ModelValidator;
 
-use Memio\Model\Contract;
-use Memio\Model\File;
-use Memio\Model\Object;
-use Memio\Validator\Constraint;
-use Memio\Validator\ConstraintValidator;
-use Memio\Validator\ModelValidator;
-use Memio\Validator\ViolationCollection;
+use Memio\Model\{
+    Contract,
+    File,
+    Object
+};
+use Memio\Validator\{
+    Constraint,
+    ConstraintValidator,
+    ModelValidator,
+    ViolationCollection
+};
 
 class FileValidator implements ModelValidator
 {
-    /**
-     * @var ContractValidator
-     */
     private $contractValidator;
-
-    /**
-     * @var ObjectValidator
-     */
     private $objectValidator;
 
-    /**
-     * @param ContractValidator $contractValidator
-     * @param ObjectValidator   $objectValidator
-     */
     public function __construct(
         ContractValidator $contractValidator,
         ObjectValidator $objectValidator
@@ -45,26 +38,17 @@ class FileValidator implements ModelValidator
         $this->constraintValidator = new ConstraintValidator();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function add(Constraint $constraint)
     {
         $this->constraintValidator->add($constraint);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($model)
+    public function supports($model) : bool
     {
         return $model instanceof File;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function validate($model)
+    public function validate($model) : ViolationCollection
     {
         if (!$this->supports($model)) {
             return new ViolationCollection();
